requirements.txt
absl-py==0.9.0
astor==0.8.1
atari-py==0.2.6
baselines==0.1.5
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
Cython==0.29.15
dill==0.3.1.1
fasteners==0.15
future==0.18.2
gast==0.2.2
glfw==1.10.1
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.1
gym==0.16.0
gym-retro==0.7.0
gym-wrappers==0.1.0
h5py==2.10.0
idna==2.8
imageio==2.6.1
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lockfile==0.12.2
Markdown==3.2.1
monotonic==1.5
mpi4py==3.0.3
# Editable install with no version control (mujoco-py==1.50.1.68)
-e /Users/gleitsmann/Python/Environments/sonic-rl-env/lib/python3.7/site-packages
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.2.0.32
opt-einsum==3.1.0
Pillow==7.0.0
progressbar2==3.47.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyglet==1.4.10
PyOpenGL==3.1.5
python-utils==2.3.0
pyzmq==18.1.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
tqdm==4.42.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
zmq==0.0.0
